Move pure unit tests out of the integration directory.
