+++ tests/integration/test_behaviors.py
@@ -1,6 +1,4 @@
 # encoding=utf-8
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 import logging
 
 import requests
