+++ tests/unit/mbtest/test_imposter.py
@@ -1,11 +1,22 @@
 # encoding=utf-8
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 import logging
 
-from mbtest.imposters import Imposter, Stub, Proxy, Response
+from hamcrest import assert_that, has_entries
+
+from mbtest.imposters import Predicate, Imposter, Stub, Response, Proxy
 
 logger = logging.getLogger(__name__)
+
+
+def test_default_predicate():
+    # Given
+    predicate = Predicate()
+
+    # When
+    structure = predicate.as_structure()
+
+    # Then
+    assert_that(structure, has_entries(caseSensitive=True, equals=has_entries()))
 
 
 def test_structure_port():
